requirements.txt
Flask~=3.1.0
pymongo~=4.11.1
python-dotenv~=1.0.1
flask-login
bcrypt~=4.3.0